# game-host/openapi/game_host_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Guess The Number API",
    "description": "API for hosting and playing the Guess The Number game",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/games/{id}": {
      "get": {
        "description": "Get the status of a game for the authenticated player",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "status": {
                      "type": "string"
                    },
                    "attempts": {
                      "type": "number"
                    },
                    "startedAt": {
                      "type": "string"
                    },
                    "finishedAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "lastGuessAt": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "status",
                    "attempts",
                    "startedAt"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/games/{id}/guess": {
      "post": {
        "description": "Submit a guess for a game for the authenticated player",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "value": {
                    "type": "number",
                    "minimum": 1,
                    "maximum": 10000
                  }
                },
                "required": [
                  "value"
                ]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "enum": [
                        "low",
                        "high",
                        "correct"
                      ]
                    },
                    "status": {
                      "type": "string"
                    },
                    "attempts": {
                      "type": "number"
                    },
                    "lastGuessAt": {
                      "type": "string"
                    },
                    "finishedAt": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "result",
                    "status",
                    "attempts",
                    "lastGuessAt"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://host.guess-the-number.com"
    }
  ]
}